requirements.txt
flask
python-dotenv
Pillow
google-genai